+++ src/format/epub/format-epub.ts
@@ -6,14 +6,16 @@
 */
 
 import { Format } from "../../config/types.ts";
-import { mergeConfigs } from "../../core/config.ts";
-import { kEPubCoverImage } from "../../config/constants.ts";
 import { ProjectConfig } from "../../project/types.ts";
+import { kEPubCoverImage } from "../../config/constants.ts";
+
+import { mergeConfigs } from "../../core/config.ts";
+
 import {
   bookConfig,
+  BookExtension,
   kBookCoverImage,
-} from "../../project/types/book/book-config.ts";
-import { BookExtension } from "../../project/types/book/book-shared.ts";
+} from "../../project/types/book/book-shared.ts";
 import { createEbookFormat } from "../formats-shared.ts";
 
 export function epubFormat(): Format {
